Complete valid string check logic
